+++ src/main/java/subway/path/ui/PathController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import subway.auth.domain.AuthenticationPrincipal;
 import subway.member.domain.LoginMember;
-import subway.path.application.PathService;
 import subway.path.dto.PathResponse;
+import subway.path.service.PathService;
 
 @RestController
 @RequestMapping("/paths")
