+++ src/main/java/subway/station/ui/StationController.java
@@ -3,9 +3,9 @@ package subway.station.ui;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.station.application.StationService;
 import subway.station.dto.StationRequest;
 import subway.station.dto.StationResponse;
+import subway.station.service.StationService;
 
 import java.net.URI;
 import java.sql.SQLException;
